Security Update: Implemented Access and Refresh Token

+++ frontend/src/components/ListSections.jsx
@@ -1,27 +1,35 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import ListCardHeading from "./ListCardHeading";
 import SimpleListCard from "./SimpleListCard"
-import { api, getToken } from "../hooks/utilityFns.jsx";
+import { api, getAccessToken } from "../hooks/utilityFns.jsx";
 import { useState } from "react";
 
-const token = getToken()
 
 // Fetch Sections
 const fetchSections = async() => {
+    const accessToken = getAccessToken()
+
+    if (!accessToken) throw new Error("No token found")
+
     const response = await api.get(`/user/cvsections`, {
         headers: {
-            Authorization: `Bearer ${token}`
+            Authorization: `Bearer ${accessToken}`
         }
     })
     
-    return response.data
+    console.log({response})
+    return response?.data
 }
 
 // Delete Section
 const deleteSection = async (sectionId) => {
+    const accessToken = getAccessToken()
+
+    if (!accessToken) throw new Error("No token found")
+
     const response = await api.delete(`/user/cvsections/${sectionId}`, {
         headers: {
-            Authorization: `Bearer ${token}`
+            Authorization: `Bearer ${accessToken}`
         }
     })
     return response.data
@@ -29,9 +37,13 @@ const deleteSection = async (sectionId) => {
 
 // Update Section
 const updateSection = async (sectionId, sectionName) => {
+    const accessToken = getAccessToken()
+
+    if (!accessToken) throw new Error("No token found")
+
     const response = await api.patch(`/user/cvsections/${sectionId}`, {sectionName}, {
         headers: {
-            Authorization: `Bearer ${token}`
+            Authorization: `Bearer ${accessToken}`
         }
     })
     return response.data
@@ -43,7 +55,7 @@ const ListSections = (props) => {
 
     const queryClient = useQueryClient()
 
-    {/* USE QUERY */}
+    {/* USE QUERY:  Fetch Sections */}
     const { data, isPending, isError, error } = useQuery({
         queryKey: ["cvSections"],
         queryFn: fetchSections
@@ -73,7 +85,11 @@ const ListSections = (props) => {
 
     }
 
-    const handleEdit = () => {
+    const handleEdit = ({title, _id}) => {
+        console.log({title, _id})
+        props.setSectionName(title)
+        props.setSectionId(_id)
+        props.setIsEditSection(true)
         props.setFormIsOpen(true)
     }
 
@@ -90,12 +106,13 @@ const ListSections = (props) => {
             {/* Display Loading */}
             {isPending && <div>Loading...</div>}
             {/* Sections List */}
+        {console.log({data})}
             <div>
                 {data?.map((section) => (
-                    <div key={section._id} className="mb-2">
+                    <div key={section?._id} className="mb-2">
                         <SimpleListCard 
-                            title = {section.sectionName}
-                            _id = {section._id}
+                            title = {section?.sectionName}
+                            _id = {section?._id}
                             handleDelete={handleDelete}
                             isLoading={isPendingDelete}
                             deleteId={deleteId}
